% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOU-bivariate-ou.R
\docType{data}
\name{bivariate_ou}
\alias{bivariate_ou}
\title{Bivariate Ornstein–Uhlenbeck Model Data}
\format{
A dataframe with 10000 rows and 4 columns
(\code{y1}, \code{y2}, \code{id}, and \code{time})
generated from the bivariate Ornstein–Uhlenbeck model
from Chow et al. (2023).
}
\usage{
bivariate_ou
}
\description{
Bivariate Ornstein–Uhlenbeck Model Data
}
\references{
Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023).
Continuous-time dynamic models:
Connections to structural equation models and other discrete-time models.
In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.).
The Guilford Press.
}
\keyword{data}
\keyword{fitOU}
