% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOU-data-ou.R
\name{DataOU}
\alias{DataOU}
\title{Preliminary Data Preparation}
\usage{
DataOU(
  data,
  observed,
  id,
  time,
  insert_na = FALSE,
  center = FALSE,
  scale = FALSE,
  scale_vars = NULL,
  initial_na = TRUE
)
}
\arguments{
\item{data}{Data frame.
A data frame object of data for potentially
multiple subjects that contain
a column of subject ID numbers
(i.e., an ID variable),
a column indicating subject-specific measurement occasions
(i.e., a TIME variable),
at least one column of observed values.}

\item{observed}{Character vector.
A vector of character strings
of the names of the observed variables in the data.}

\item{id}{Character string.
A character string of the name of the ID variable in the data.}

\item{time}{Character string.
A character string of the name of the TIME variable in the data.}

\item{insert_na}{Logical.
Insert \code{NA} to \code{observed} variables for existing \code{time} points.}

\item{center}{Logical.
If \code{center = TRUE}, mean center by \code{id}.}

\item{scale}{Logical.
If \code{scale = TRUE}, standardize by \code{id}.}

\item{scale_vars}{Character vector.
A vector of character strings
of the names of the observed variables to center/scale.}

\item{initial_na}{Logical.
Iteratively remove rows where any observed variable
for the first time point has \code{NA}.}
}
\description{
Preliminary Data Preparation
}
\examples{
data <- DataOU(
  data = bivariate_ou,
  observed = c("y1", "y2"),
  id = "id",
  time = "time"
)
summary(data)

}
\seealso{
Other Fit Ornstein–Uhlenbeck Model Functions: 
\code{\link{FitOU}()},
\code{\link{ModelOU}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fit Ornstein–Uhlenbeck Model Functions}
\keyword{fit}
\keyword{fitOU}
