% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOU-package.R
\docType{package}
\name{fitOU-package}
\alias{fitOU-package}
\alias{_PACKAGE}
\title{fitOU: The Ornstein–Uhlenbeck Model}
\description{
Fit the Ornstein–Uhlenbeck model using the 'dynr' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijapesigan/fitOU}
  \item \url{https://ijapesigan.github.io/fitOU/}
  \item Report bugs at \url{https://github.com/ijapesigan/fitOU/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.ijapesigan@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
