% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOU-fit-ou.R
\name{FitOU}
\alias{FitOU}
\title{Fit the Ornstein–Uhlenbeck Model}
\usage{
FitOU(
  data,
  observed,
  id,
  time,
  mu0 = NULL,
  sigma0 = NULL,
  mu_start = NULL,
  phi_start = NULL,
  sigma_start = NULL,
  theta_start = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame.
A data frame object of data for potentially
multiple subjects that contain
a column of subject ID numbers
(i.e., an ID variable),
a column indicating subject-specific measurement occasions
(i.e., a TIME variable),
at least one column of observed values.}

\item{observed}{Character vector.
A vector of character strings
of the names of the observed variables in the data.}

\item{id}{Character string.
A character string of the name of the ID variable in the data.}

\item{time}{Character string.
A character string of the name of the TIME variable in the data.}

\item{mu0}{Numeric vector.
Mean of initial latent variable values
(\eqn{\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}}).
If \code{mu0 = NULL}, a vector of zeros is used.}

\item{sigma0}{Numeric matrix.
Covariance matrix
of initial latent variable values
(\eqn{\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}}).
If \code{sigma0 = NULL}, an identity matrix is used.}

\item{mu_start}{Numeric vector.
Starting values of the \code{mu} vector,
that is,
the long-term mean or equilibrium level.
If \code{mu_start = NULL},
a vector means of the observed variables is used.}

\item{phi_start}{Numeric matrx.
Starting values of the \code{phi} matrix,
that is,
the rate of mean reversion,
determining how quickly the variable returns to its mean.
If \code{phi_start = NULL}, a matrix of zeros is used.}

\item{sigma_start}{Numeric matrx.
Starting values of the \code{sigma} matrix,
that is,
the matrix of volatility
or randomness in the process.
If \code{sigma_start = NULL}, an identity matrix is used.}

\item{theta_start}{Numeric matrx.
Starting values of the \code{theta} matrix,
that is,
the measurement error covariance matrix
(\eqn{\boldsymbol{\Theta}}).
If \code{theta_start = NULL}, an identity matrix is used.}

\item{...}{Additional arguments to pass to \code{\link[dynr:dynr.cook]{dynr::dynr.cook()}}.}
}
\description{
This is a wrapper function that makes fitting the Ornstein–Uhlenbeck model
convenient using the \code{dynr} package.
}
\details{
The measurement model is given by
\deqn{
  \mathbf{y}_{i, t}
  =
  \boldsymbol{\nu}
  +
  \boldsymbol{\Lambda}
  \boldsymbol{\eta}_{i, t}
  +
  \boldsymbol{\varepsilon}_{i, t}
  \quad
  \mathrm{with}
  \quad
  \boldsymbol{\varepsilon}_{i, t}
  \sim
  \mathcal{N}
  \left(
  \mathbf{0},
  \boldsymbol{\Theta}
  \right)
}

where \eqn{\mathbf{y}_{i, t}}, \eqn{\boldsymbol{\eta}_{i, t}},
and \eqn{\boldsymbol{\varepsilon}_{i, t}}
are random variables and \eqn{\boldsymbol{\nu}},
\eqn{\boldsymbol{\Lambda}},
and \eqn{\boldsymbol{\Theta}} are model parameters.
\eqn{\mathbf{y}_{i, t}} is a vector of observed random variables
at time \eqn{t} and individual \eqn{i},
\eqn{\boldsymbol{\eta}_{i, t}} is a vector of latent random variables
at time \eqn{t} and individual \eqn{i},
and \eqn{\boldsymbol{\varepsilon}_{i, t}}
is a vector of random measurement errors
at time \eqn{t} and individual \eqn{i},
while \eqn{\boldsymbol{\nu}} is a vector of intercept,
\eqn{\boldsymbol{\Lambda}} is a matrix of factor loadings,
and \eqn{\boldsymbol{\Theta}} is the covariance matrix of
\eqn{\boldsymbol{\varepsilon}}.

The dynamic structure is given by

\deqn{
  \mathrm{d} \boldsymbol{\eta}_{i, t}
  =
  \boldsymbol{\Phi}
  \left(
  \boldsymbol{\mu}
  -
  \boldsymbol{\eta}_{i, t}
  \right)
  \mathrm{d}t
  +
  \boldsymbol{\Sigma}^{\frac{1}{2}}
  \mathrm{d}
  \mathbf{W}_{i, t}
}

where \eqn{\boldsymbol{\mu}} is the long-term mean or equilibrium level,
\eqn{\boldsymbol{\Phi}} is the rate of mean reversion,
determining how quickly the variable returns to its mean,
\eqn{\boldsymbol{\Sigma}} is the matrix of volatility
or randomness in the process, and \eqn{\mathrm{d}\boldsymbol{W}}
is a Wiener process or Brownian motion,
which represents random fluctuations.
}
\examples{
\dontrun{
FitOU(
  data = bivariate_ou,
  observed = c("y1", "y2"),
  id = "id",
  time = "time",
  verbose = FALSE
)
}
}
\references{
Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023).
Continuous-time dynamic models:
Connections to structural equation models and other discrete-time models.
In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.).
The Guilford Press.

Ou, L., Hunter, M. D., & Chow, S.-M. (2019).
What's for dynr:
A package for linear and nonlinear dynamic modeling in R.
\emph{The R Journal}, \emph{11}(1), 91.
\doi{10.32614/rj-2019-012}

Uhlenbeck, G. E., & Ornstein, L. S. (1930).
On the theory of the brownian motion.
\emph{Physical Review}, \emph{36}(5), 823–841.
\doi{doi.org/10.1103/physrev.36.823}
}
\seealso{
Other Fit Ornstein–Uhlenbeck Model Functions: 
\code{\link{FitOUID}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fit Ornstein–Uhlenbeck Model Functions}
\keyword{fit}
\keyword{fitOU}
